+++ packages/phishing-controller/src/PhishingController.ts
@@ -1466,6 +1466,7 @@ export class PhishingController extends BaseController<
       return {
         result_type: AddressScanResultType.ErrorResult,
         label: '',
+        address_alert_response_flagged_by: [],
       };
     }
 
@@ -1477,6 +1478,7 @@ export class PhishingController extends BaseController<
       return {
         result_type: AddressScanResultType.ErrorResult,
         label: '',
+        address_alert_response_flagged_by: [],
       };
     }
 
@@ -1486,6 +1488,8 @@ export class PhishingController extends BaseController<
       return {
         result_type: cachedResult.result_type,
         label: cachedResult.label,
+        address_alert_response_flagged_by:
+          cachedResult.address_alert_response_flagged_by ?? [],
       };
     }
 
@@ -1521,18 +1525,22 @@ export class PhishingController extends BaseController<
       return {
         result_type: AddressScanResultType.ErrorResult,
         label: '',
+        address_alert_response_flagged_by: [],
       };
     } else if ((apiResponse as { error?: string }).error) {
       return {
         result_type: AddressScanResultType.ErrorResult,
         label: '',
+        address_alert_response_flagged_by: [],
       };
     }
 
     const scanResult = apiResponse as AddressScanResult;
+    const flaggedBy = scanResult.address_alert_response_flagged_by ?? [];
     const result: AddressScanCacheData = {
       result_type: scanResult.result_type,
       label: scanResult.label,
+      address_alert_response_flagged_by: flaggedBy,
     };
 
     this.#addressScanCache.set(cacheKey, result);
@@ -1540,6 +1548,7 @@ export class PhishingController extends BaseController<
     return {
       result_type: scanResult.result_type,
       label: scanResult.label,
+      address_alert_response_flagged_by: flaggedBy,
     };
   }